Порт геодаты с легиона и ЛК Включаем динамический лос Кое какие ренеймы и унификация Обновление поиска пути Доп правки для того чтоб все это завелось

+++ src/server/shared/Memory.h
@@ -3,15 +3,17 @@
 #ifndef _MEMORY_H
 #define _MEMORY_H
 
+#include "DetourAlloc.h"
+
 //  memory management
-inline void* dtCustomAlloc(int size, dtAllocHint /*hint*/)
+inline void* dtCustomAlloc(size_t size, dtAllocHint /*hint*/)
 {
     return (void*)new unsigned char[size];
 }
 
 inline void dtCustomFree(void* ptr)
 {
-    delete [] (unsigned char*)ptr;
+    delete[](unsigned char*)ptr;
 }
 
 #endif
